Switch implementation from IAM Federated Access to AWS SSO
