+++ README.md
@@ -1,11 +1,11 @@
-# modernisation-platform-terraform-iam-federated-access
+# moj-terraform-aws-sso
 
-This repository holds a Terraform module that creates an Auth0 application and associated integrations to enable IAM Federated Access in AWS.
+This repository holds a Terraform module that creates an Auth0 application and associated integrations to enable AWS SSO.
 
 ## Usage
 ```
-module "iam_federation" {
-  source              = "github.com/ministryofjustice/modernisation-platform-terraform-iam-federated-access"
+module "sso" {
+  source                     = "github.com/ministryofjustice/moj-terraform-aws-sso"
   auth0_tenant_domain        = ""
   auth0_client_id            = ""
   auth0_client_secret        = ""
@@ -13,25 +13,22 @@ module "iam_federation" {
   auth0_github_client_id     = ""
   auth0_github_client_secret = ""
   auth0_github_allowed_orgs  = ["ministryofjustice"]
-
+  auth0_allowed_domains      = ["@digital.justice.gov.uk"]
+  auth0_aws_sso_acs_url      = "https://region.signin.aws.amazon.com/platform/saml/acs/${random-key}"
+  auth0_aws_sso_issuer_url   = "https://region.signin.aws.amazon.com/platform/saml/${random-key}"
 }
 ```
 
 ## Inputs
-| Name                       | Description                                                         | Type    | Default | Required |
-|----------------------------|---------------------------------------------------------------------|---------|---------|----------|
-| auth0_tenant_domain        | Tenant domain from the Auth0 account to create applicable resources | string  | n/a     | yes      |
-| auth0_client_id            | Auth0 application (Machine to Machine) client ID to utilise         | string  | n/a     | yes      |
-| auth0_client_secret        | Auth0 application (Machine to Machine) client secret to utilise     | string  | n/a     | yes      |
-| auth0_debug                | Whether to turn Auth0 debugging on or off                           | boolean | `false` | no       |
-| auth0_github_client_id     | GitHub OAuth app client ID for an Auth0 social connection           | string  | n/a     | yes      |
-| auth0_github_client_secret | GitHub OAuth app client secret for an Auth0 social connection       | string  | n/a     | yes      |
-| auth0_github_allowed_orgs  | A list of organisations a user has to be part of to authenticate    | list    | n/a     | yes      |
-
-## Outputs
-| Name            | Description                            | Sensitive |
-|-----------------|----------------------------------------|-----------|
-| saml_login_page | URL for the SAML login page from Auth0 | no        |
-
-## Looking for issues?
-If you're looking to raise an issue with this module, please create a new issue in the [Modernisation Platform repository](https://github.com/ministryofjustice/modernisation-platform/issues).
+| Name                       | Description                                                          | Type    | Default | Required |
+|----------------------------|----------------------------------------------------------------------|---------|---------|----------|
+| auth0_tenant_domain        | Tenant domain from the Auth0 account to create applicable resources  | string  | n/a     | yes      |
+| auth0_client_id            | Auth0 application (Machine to Machine) client ID to utilise          | string  | n/a     | yes      |
+| auth0_client_secret        | Auth0 application (Machine to Machine) client secret to utilise      | string  | n/a     | yes      |
+| auth0_debug                | Whether to turn Auth0 debugging on or off                            | boolean | `false` | no       |
+| auth0_github_client_id     | GitHub OAuth app client ID for an Auth0 social connection            | string  | n/a     | yes      |
+| auth0_github_client_secret | GitHub OAuth app client secret for an Auth0 social connection        | string  | n/a     | yes      |
+| auth0_github_allowed_orgs  | A list of organisations a user has to be part of to authenticate     | list    | n/a     | yes      |
+| auth0_allowed_domains      | A list of authorised domains a user can have in their GitHub account | list    | n/a     | yes      |
+| auth0_aws_sso_acs_url      | AWS SSO ACS URL, as provided by AWS when you set up AWS SSO          | string  | n/a     | yes      |
+| auth0_aws_sso_issuer_url   | AWS SSO Issuer URL, as provided by AWS when you set up AWS SSO       | string  | n/a     | yes      |
