--- outputs.tf
@@ -1,3 +1,0 @@
-output "saml_login_page" {
-  value = "https://${var.auth0_tenant_domain}/samlp/${auth0_client.saml.client_id}"
-}
