+++ variables.tf
@@ -33,3 +33,18 @@ variable "auth0_github_allowed_orgs" {
   description = "A list of GitHub organisations a user has to be part of"
   type        = list(string)
 }
+
+variable "auth0_allowed_domains" {
+  description = "A list of authorised domains a user must have as part of their GitHub email addresses"
+  type        = list(string)
+}
+
+variable "auth0_aws_sso_acs_url" {
+  description = "AWS SSO: ACS URL"
+  type        = string
+}
+
+variable "auth0_aws_sso_issuer_url" {
+  description = "AWS SSO: Issuer URL"
+  type        = string
+}
