+++ versions.tf
@@ -7,8 +7,5 @@ terraform {
     aws = {
       source = "hashicorp/aws"
     }
-    external = {
-      source = "hashicorp/external"
-    }
   }
 }
